dom/base/test/test_bug390735.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=390735
-->
<head>
  <meta http-equiv="content-type" content="text/html; charset=utf-8">
  <title>Test for Bug 390735</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">
/** Test for Bug 390735 **/

var contents = document.getElementsByTagName("head")[0].innerHTML;
var expectedFind = "<meta http-equiv=\"content-type\" content=\"text/html; charset=utf-8\">";

ok(contents.indexOf(expectedFind) > -1, "The meta tag element was not found");
</script>
</pre>
</body>
</html>
